Refactor, send email with confirmation codes
Added composer, a PHP package manager of sorts to manage our new
dependency on the PHPMailer library.  As a result, the dockerfile
has to be edited to accomodate composer.

In addition to adding composer, the dockerfile has been edited such
that the configuration files for nginx and the various php.ini's
have been moved to the "config" subdirectory.  Also, the volume for
the php.ini's has been removed from docker-compose.yml; instead they
will be copied into the image during "docker compose build".

Application configuration has partially moved out of docker-compose.yml
into a new file called "bb_config.ini".  This is also where the SMTP
settings live.  Note that this file is deliberately not included in
the repository; instead there is "bb_config.ini.example" which should
be copied without the "example" suffix and configured appropriately.

The confirmation code for new registrants has been made a 32-bit
integer.  Ideally it would be longer, but the size of PHP integers
machine-dependent.